simplify introspector
