+++ lib/introspector/example/tree-sitter-parser.ts
@@ -1,6 +1,6 @@
 import { parserFromWasm } from "deno-tree-sitter/main.js";
 import typescript from "common-tree-sitter-languages/typescript.js";
-import { defaultClassTreeSitterCaptureNameMap } from "../from-typescript-class.ts";
+import { defaultClassTreeSitterCaptureNameMap } from "#/lib/introspector/tree-sitter-introspector.ts";
 
 /**
  * EXAMPLE_TYPESCRIPT_TREE_SITTER_PARSER is a tree-sitter parser for TypeScript.
