+++ lib/introspector/introspector.ts
@@ -1,0 +1,26 @@
+/**
+ * Introspector is an interface for TypeScript type introspection.
+ */
+export interface Introspector {
+  /**
+   * introspectClass introspects a TypeScript class's properties.
+   */
+  introspectClass(sourceCode: string, identifier: string): Introspection;
+}
+
+/**
+ * Introspection is a TypeScript class introspection.
+ */
+export interface Introspection {
+  extends: string[];
+  properties: IntrospectedProperty[];
+}
+
+/**
+ * IntrospectedProperty is a TypeScript class property introspection.
+ */
+export interface IntrospectedProperty {
+  name: string;
+  type: string;
+  optional: boolean;
+}
